codestral-mamba/pyproject.toml
[tool.poetry]
name = "codestral-mamba"
version = "0.1.0"
description = ""
authors = ["Elvin Gomez <elvin@codingmindset.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
mistralai = "^0.4.2"
gradio = "^4.38.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
